Improve error logging when file load fails
We were faced with the following error on startup:
`File with name config.nd does not exist in storage OSystem`. 
It turns out that the log is not the true resolved filename that orient looks up.
It would be helpful to figure out what file is missing if we have the true filename.